fix(assets): rename inmemory_assets.py to test_inmemory_assets.py so pytest discovers it
